+++ lib/comeonin/pbkdf2.ex
@@ -45,21 +45,27 @@ defmodule Comeonin.Pbkdf2 do
 
   @doc """
   Hash the password with a salt which is randomly generated.
+
+  To change the complexity (and the time taken) of the  password hash
+  calculation, you need to change the value for `pbkdf2_rounds`
+  in the config file.
   """
   def hashpwsalt(password) do
-    hashpass(password, gen_salt)
+    hashpass(password, gen_salt, Config.pbkdf2_rounds)
   end
 
   @doc """
-  This function first checks that the password is long enough and
-  contains at least one number and one punctuation character. The
-  password is then hashed only if the password is considered strong
-  enough. Read the docs for the Comeonin.Password module for more information.
+  This is a convenience function that checks the strength of a password
+  before hashing it. The password is then hashed only if the password is
+  considered strong enough. For more details about password strength,
+  read the documentation for the Comeonin.Password module.
   """
-  def hashpwsalt(password, true) do
-    Password.valid_password?(password) and hashpass(password, gen_salt)
+  def signup_user(password) do
+    case Password.valid_password?(password) do
+      true -> {:ok, hashpwsalt(password)}
+      message -> {:error, message}
+    end
   end
-  def hashpwsalt(password, _), do: hashpass(password, gen_salt)
 
   defp format(hash, salt, rounds) do
     "$pbkdf2-sha512$#{rounds}$#{Pbkdf2Base64.encode(salt)}$#{Pbkdf2Base64.encode(hash)}"